refactor: decouple tree demos from core implementation
